Integrating the Factory & Dependency Inversion Design Patterns
The Factory pattern is easy and basically working fine. (justneeds to be integrated.) but the DI pattern is still a little fuckie. The ground work is there but the connections between all the different views and windows arent setup (ha! not even figured out) yet. The reliability is not nearly good enough. any little problem causes object null ecxeptions.